+++ src/test/java/com/acceptance/LoginPage.java
@@ -1,8 +1,7 @@
-package atda;
+package com.acceptance;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.support.ui.ExpectedConditions;
 
 public class LoginPage extends BasePage {
