+++ src/test/java/com/acceptance/ProductsPage.java
@@ -1,4 +1,4 @@
-package atda;
+package com.acceptance;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
